Fix sync/async mock mismatch
Members of mocks will inherit their "function color" (sync or async)
from their parent. Objects returned from asynchronous code might have
synchronous methods themselves, but the mock classes need to be told
about it if they are used in the tested code, otherwise it will either
raise an exception (awaiting the result of a sync callable) or raise a
warning in tests when the awaitable result of the mock isn’t awaited.

Signed-off-by: Nils Philippsen <nils@redhat.com>